pyproject.toml
[project]
name = "ragster"
requires-python = ">=3.10,<3.12"
version = "0.1.0"
dependencies = [
    "airbyte>=0.23.0",
    "chromadb>=0.6.3",
    "dagster-components",
    "duckdb>=1.2.0",
]

[project.entry-points]
"dagster.components" = { ragster = "ragster.lib"}

[dependency-groups]
dev = [
    "dagster-webserver",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.dagster]
module_name = "ragster.definitions"
code_location_name = "ragster"

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "ragster"

[tool.setuptools.packages.find]
exclude=["ragster_tests"]


